Move litegraph utils
