+++ browser_tests/groupNode.spec.ts
@@ -1,9 +1,6 @@
 import { expect } from '@playwright/test'
-import {
-  ComfyPage,
-  NodeReference,
-  comfyPageFixture as test
-} from './fixtures/ComfyPage'
+import { ComfyPage, comfyPageFixture as test } from './fixtures/ComfyPage'
+import type { NodeReference } from './fixtures/utils/litegraphUtils'
 
 test.describe('Group Node', () => {
   test.afterEach(async ({ comfyPage }) => {
